requirements.txt
Django==5.2.5
djangorestframework
django-phonenumber-field
phonenumbers